fix(constants): rename folder
