+++ src/utils/constants.js
@@ -1,1 +1,2 @@
+export const BASE_URL = 'https://kinder.acceleratorpracticum.ru/';
 export const EMAIL_PATTERN = '[a-zA-Z0-9_.]+@[a-zA-Z0-9_]+\\.[a-z]{2,}';
